Added NewTodoServlet, DeleteTodoServlet refactoring LoginServlet, changed entity join add newTodo, delete todo from db works (after deleting page doesn't show updated form)

+++ src/main/java/com/fruitella/todo/DAO/UserDAO.java
@@ -10,9 +10,9 @@ public interface UserDAO {
 
     void updateUser(Users user);
 
-    void deleteUser(int userId);
+    void deleteUser(long userId);
 
-    Users getUserById(int userId);
+    Users getUserById(long userId);
 
     Users getUserByUsername(String username);
